feat: add data aggregation application
