+++ packages/vmind/src/applications/dataAggregation/prompts/GPT/index.ts
@@ -1,0 +1,11 @@
+import { Prompt } from 'src/base/tools/prompt';
+import { getQueryDatasetPrompt } from './template';
+import { DataAggregationContext } from '../../types';
+
+export class GPTDataAggregationPrompt extends Prompt<DataAggregationContext> {
+  getPrompt(context: DataAggregationContext) {
+    const { llmOptions } = context;
+    const QueryDatasetPrompt = getQueryDatasetPrompt(llmOptions.showThoughts ?? true);
+    return QueryDatasetPrompt;
+  }
+}
