+++ packages/vmind/src/applications/dataAggregation/taskNodes/executeQuery/index.ts
@@ -1,0 +1,4 @@
+import { RuleBasedTaskNode } from 'src/base/taskNode/ruleBasedTaskNode';
+import { Transformer } from 'src/base/tools/transformer';
+
+export const ExecuteQueryTaskNode = new RuleBasedTaskNode();
