+++ packages/vmind/src/applications/dataAggregation/types.ts
@@ -1,0 +1,9 @@
+import { DataItem, ILLMOptions, SimpleFieldInfo } from 'src/typings';
+
+export type SQL = string;
+export type DataAggregationContext = {
+  llmOptions: ILLMOptions;
+  userInput: string;
+  fieldInfo: SimpleFieldInfo[];
+  sourceDataset: DataItem[];
+};
