+++ packages/vmind/src/base/application/index.ts
@@ -8,20 +8,38 @@ import { IApplication } from './types';
  * Application can be seen as a collection of a series of TaskNodes. Applications can reference each other (equivalent to reusing TaskNodes to complete tasks)
  */
 export class BaseApplication<Context, DSL> implements IApplication<Context, DSL> {
-  tasks: BaseTaskNode<Context, any>[];
+  tasks: {
+    name: string;
+    task: BaseTaskNode<Context, any>;
+  }[];
   context: Context;
   chatManager: ChatManager;
-  constructor(tasks: BaseTaskNode<Context, any>[], context: Context) {
+  constructor(
+    tasks: {
+      name: string;
+      task: BaseTaskNode<Context, any>;
+    }[],
+    context: Context
+  ) {
     this.tasks = tasks;
     this.context = context;
     this.chatManager = new ChatManager();
   }
 
   async runTasks() {
-    const result: DSL = this.tasks.reduce(async (pre: any, task: BaseTaskNode<Context, any>) => {
-      const result = await task.executeTask(pre);
-      return result;
-    }, this.context);
+    const result: DSL = this.tasks.reduce(
+      async (
+        pre: any,
+        curTask: {
+          name: string;
+          task: BaseTaskNode<Context, any>;
+        }
+      ) => {
+        const result = await curTask.task.executeTask(pre);
+        return result;
+      },
+      this.context
+    );
     return result;
   }
