+++ packages/vmind/src/base/tools/parser/types.ts
@@ -1,6 +1,6 @@
 import { Transformer } from '../transformer';
 
 export interface IParser<DSL> {
-  transformer: Transformer<string, unknown, DSL>;
+  transformer: Transformer<string, null, DSL>;
   parse: (input: string) => DSL;
 }
