+++ packages/vmind/src/base/tools/transformer/index.ts
@@ -6,7 +6,11 @@ import { ITransformer } from './types';
  * The subclass needs to rewrite the transform method to complete the specific data conversion
  */
 export class Transformer<Input, Context, DSL> implements ITransformer<Input, Context, DSL> {
+  transformFunc: (input: Input, context: Context) => DSL;
+  constructor(transFunc: (input: Input, context: Context) => DSL) {
+    this.transformFunc = transFunc;
+  }
   transform(input: Input, context: Context): DSL {
-    return input as unknown as DSL;
+    return this.transformFunc(input, context);
   }
 }
