+++ packages/vmind/src/base/tools/transformer/types.ts
@@ -1,3 +1,4 @@
 export interface ITransformer<Input, Context, DSL> {
+  transformFunc: (input: Input, context: Context) => DSL;
   transform: (input: Input, context: Context) => DSL;
 }
